pyproject.toml
[project]
name = "bottelegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pandas>=2.3.0",
    "playwright>=1.52.0",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "tabulate>=0.9.0",
]
